update sdl & spark version

+++ src/main/scala/com/sample/ReduceNycCSVTransformer.scala
@@ -6,6 +6,6 @@ import org.apache.spark.sql.{DataFrame, SparkSession}
 class ReduceNycCSVTransformer extends CustomDfTransformer {
   override def transform(session: SparkSession, options: Map[String,String], df: DataFrame, dataObjectId: String) : DataFrame = {
     import session.implicits._
-    df.select($"id", $"name")
+    df.select($"id".cast("int"), $"name")
   }
 }
